requirements.txt
PyQt5==5.15.9
keyboard==0.13.5
Pillow==10.0.0
openai==1.3.0
python-dotenv==1.0.0 